更新 'SrcJuDataCache.js'

+++ SrcJuDataCache.js
@@ -2,7 +2,9 @@ function cacheData(data){
     let jkdata = data;
     let cfilename = jkdata.name+'-'+jkdata.type;
     let cachefile = `hiker://files/cache/_fileSelect_${cfilename}.json`;
-    writeFile(cachefile,JSON.stringify(jkdata));
+    if (!fileExist(cachefile)) {
+        writeFile(cachefile,JSON.stringify(jkdata));
+    }
 }
 
 function readData(cfilename,datatype){
